address request changes and refactor code
